@@ -1,46 +1,58 @@
 {
   "homepage": "https://vitalpointai.com",
-  "name": "personas",
+  "name": "catalyst",
   "version": "0.0.1",
+  "authors": [
+    "Aaron Luhning, <a.luhning@vitalpoint.ai"
+  ],
   "private": true,
   "dependencies": {
+    "@ceramicnetwork/http-client": "^0.9.6",
+    "@ceramicstudio/idx": "^0.8.0",
+    "@ceramicstudio/idx-tools": "^0.8.0",
     "@ledgerhq/hw-transport-u2f": "^5.34.0",
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.56",
     "@sentry/browser": "^5.28.0",
     "@textile/hub": "^4.1.0",
+    "3id-connect": "^0.1.0",
+    "3id-did-provider": "^0.5.1",
+    "aguid": "^2.0.0",
     "axios": "^0.21.1",
     "big.js": "^5.2.2",
+    "bip39": "^2.4.0",
     "connected-react-router": "^6.6.0",
     "env-cmd": "^10.1.0",
     "fetch-send-json": "0.0.2",
     "fontsource-roboto": "^4.0.0",
     "fs": "0.0.1-security",
+    "ipfs-http-client": "^49.0.4",
     "js-sha256": "^0.9.0",
     "keccak256": "^1.0.2",
     "lodash": "^4.17.20",
-    "mixpanel-browser": "^2.40.0",
+    "mixpanel-browser": "^2.41.0",
     "near": "^0.2.2",
-    "near-api-js": "^0.36.0",
+    "near-api-js": "^0.39.0",
     "near-ledger-js": "^0.1.2",
-    "near-seed-phrase": "^0.1.0",
+    "near-seed-phrase": "^0.1.1",
+    "near-web3-provider": "^1.2.0",
     "prop-types": "^15.7.2",
     "query-string": "^6.13.6",
-    "react": "^16.6.3",
+    "react": "^16.13.1",
     "react-beautiful-dnd": "^13.0.0",
     "react-charts": "^2.0.0-beta.7",
     "react-datepicker": "^3.3.0",
     "react-dnd": "^11.1.3",
     "react-dnd-html5-backend": "^11.1.3",
     "react-dom": "^16.13.1",
-    "react-hook-form": "^6.3.3",
+    "react-hook-form": "^6.15.4",
     "react-images-upload": "^1.2.8",
     "react-localize-redux": "^3.5.3",
-    "react-phone-number-input": "2.3.11",
+    "react-phone-number-input": "^3.1.16",
     "react-qr-svg": "^2.2.2",
     "react-quill": "^1.3.5",
-    "react-redux": "^7.2.2",
+    "react-redux": "^7.2.0",
     "react-router-dom": "^5.2.0",
     "reduce-reducers": "^1.0.4",
     "redux": "^4.0.5",
@@ -70,37 +83,39 @@
     "build-storybook": "build-storybook"
   },
   "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
+    "last 2 Chrome versions"
   ],
+  "resolutions": {
+    "@babel/preset-env": "7.13.8",
+    "cids": "1.1.6"
+  },
   "devDependencies": {
     "@babel/core": "^7.12.9",
-    "@babel/plugin-proposal-decorators": "^7.12.1",
+    "@babel/plugin-proposal-class-properties": "^7.13.0",
+    "@babel/plugin-proposal-decorators": "^7.13.5",
     "@babel/preset-env": "^7.12.7",
     "@babel/preset-react": "^7.12.7",
-    "@sentry/cli": "^1.60.1",
-    "@storybook/addon-actions": "^6.0.28",
-    "@storybook/addon-essentials": "^6.1.9",
-    "@storybook/addon-links": "^6.1.9",
+    "@sentry/cli": "^1.62.0",
+    "@storybook/addon-actions": "^6.1.12",
+    "@storybook/addon-essentials": "^6.1.12",
+    "@storybook/addon-links": "^6.1.12",
     "@storybook/react": "^6.1.10",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^26.6.1",
     "babel-loader": "^8.1.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.5",
-    "core-js": "^3.8.1",
+    "core-js": "^3.9.1",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.15.5",
-    "eslint": "5.16.0",
-    "eslint-config-react-app": "^3.0.8",
+    "eslint": "7.21.0",
+    "eslint-config-react-app": "^6.0.0",
     "eslint-plugin-flowtype": "^5.1.1",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.21.5",
     "jest": "^26.6.1",
-    "parcel-bundler": "^1.12.4",
-    "parcel-plugin-bundle-visualiser": "^1.2.0",
+    "parcel-bundler": "^1.3.1",
+    "parcel-plugin-bundle-visualiser": "^0.8.2",
     "react-is": "^16.13.1"
   }
 }
