@@ -57,7 +57,8 @@ export default createGlobalStyle`
     }
 
     body, p, h1, h2, h3, h4, h5, h6, button, .button, input, select, textarea {
-        font-family: "Inter", Lato, "Lucida Grande", Tahoma, Sans-Serif !important;
+     //   font-family: "Inter", Lato, "Lucida Grande", Tahoma, Sans-Serif !important;
+     font-family: 'Roboto', sans-serif;
     }
 
     h1, .h1 {
