@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from 'react'
+import { useParams } from 'react-router-dom'
 import LogoutButton from '../common/LogoutButton/logoutButton'
 import ActionSelector from '../ActionSelector/actionSelector'
 import ProposalList from '../ProposalList/proposalList'
@@ -6,6 +7,12 @@ import BalanceChart from '../BalanceGraphs/balanceGraph'
 import RightSideDrawer from './RightSideDrawer'
 import InfoPopup from '../../../common/InfoPopup'
 import { Translate } from 'react-localize-redux'
+import { IDX } from '@ceramicstudio/idx'
+
+import { dao } from '../../../../utils/dao'
+import { dids } from '../../../../utils/dids'
+import { ceramic } from '../../../../utils/ceramic'
+import { wallet } from '../../../../utils/wallet'
 
 // Material UI imports
 import { makeStyles } from '@material-ui/core/styles'
@@ -103,80 +110,367 @@ export default function AppFramework(props) {
     const [guildBalanceChip, setGuildBalanceChip] = useState()
     const [escrowBalanceChip, setEscrowBalanceChip] = useState()
     const [nearPrice, setNearPrice] = useState()
+
+    const [tabValue, setTabValue] = useState('1')
+    const [contract, setContract] = useState()
+    const [allMemberInfo, setAllMemberInfo] = useState([])
+    const [contractIdx, setContractIdx] = useState()
+    const [didsContract, setDidsContract] = useState()
+    const [memberStatus, setMemberStatus] = useState()
+    const [memberInfo, setMemberInfo] = useState()
+    const [currentPeriod, setCurrentPeriod] = useState()
+    const [curUserIdx, setCurUserIdx] = useState()
+    const [curUserCeramicClient, setCurUserCeramicClient] = useState()
+    const [accountId, setAccountId] = useState()
+    const [summoner, setSummoner] = useState()
+    const [totalShares, setTotalShares] = useState()
+    const [escrowBalance, setEscrowBalance] = useState()
+    const [guildBalance, setGuildBalance] = useState()
+    const [depositToken, setDepositToken] = useState()
+    const [proposalDeposit, setProposalDeposit] = useState()
+    const [periodDuration, setPeriodDuration] = useState()
+    const [proposalEvents, setProposalEvents] = useState([])
     
     const classes = useStyles()
     
     const {      
-      tabValue,
-      handleTabValueState, 
-      accountId,
-      memberStatus,
-      memberInfo,
-      depositToken,
+    //  handleSetCurrentPeriod,
+     
+     
+    //  memberStatus,
+    //  memberInfo,
+    //  depositToken,
       tributeToken,
       tributeOffer,
       processingReward,
-      proposalDeposit,
-      guildBalance,
-      escrowBalance,
-      proposalEvents,
+     // proposalDeposit,
+     // guildBalance,
+     // escrowBalance,
+     // proposalEvents,
       handleProposalEventChange,
-      handleGuildBalanceChanges,
-      handleEscrowBalanceChanges,
+     // handleGuildBalanceChanges,
+     // handleEscrowBalanceChanges,
       handleUserBalanceChanges,
-      currentPeriod,
-      periodDuration,
+      handleSuccessMessage,
+      handleErrorMessage,
+      handleSnackBarOpen,
+      handleHasDao,
+      hasDao,
+      factoryContract,
+    //  currentPeriod,
+    //  periodDuration,
       tokenName,
       minSharePrice,
       proposalComments,
-      contract,
+     
       daoContract,
       handleInitChange,
-      summoner,
-      totalShares,
-      allMemberInfo } = props
+      appIdx,
+     // summoner,
+     // totalShares,
 
+    // accountId,
+    handleContractIdx,
+     appClient,
+     refreshAccount,
+    //  idx,
+     // didsContract,
+     // contractId,
+     // contractIdx 
+    } = props
+      
+
+      const {
+        contractId
+      } = useParams()
+      console.log('contractid', contractId)
+
+      async function handleGuildBalanceChanges() {
+        try {
+          let currentGuildBalance = await contract.getGuildTokenBalances()
+          if(currentGuildBalance) {
+            setGuildBalance(currentGuildBalance)
+            await refreshAccount()
+          }
+          return true
+        } catch (err) {
+          return false
+        }
+      }
+
+      async function handleEscrowBalanceChanges() {
+        try {
+          let currentEscrowBalance = await contract.getEscrowTokenBalances()
+          if(currentEscrowBalance) {
+            setEscrowBalance(currentEscrowBalance)
+            await refreshAccount()
+          }
+          return true
+        } catch (err) {
+          return false
+        }
+      }
+      
+      function handleTabValueState(value) {
+        setTabValue(value)
+      }
+    
       useEffect(
         () => {
           let isMounted = true; // note this flag denote mount status
+
             async function fetchData() {
-              if(memberStatus && memberInfo !== undefined) {
+
+              let accountObj = await dao.loadAccountObject()
+              let accountId     
+              if(accountObj){
+                  accountId = accountObj.accountId
+                  setAccountId(accountId)
+              }
+              let curAccount = await wallet.getAccount(accountId)
+              console.log('curaccount', curAccount)
+              let seed = await ceramic.getSeed(curAccount)
+              let thisCurrentUserCeramicClient = await ceramic.getCeramic(curAccount, seed)
+              setCurUserCeramicClient(thisCurrentUserCeramicClient)
+              
+              //Set App Ceramic Client
+              let appSeed = Buffer.from(process.env.FACTORY_PRIV_KEY.slice(0, 32))
+              let appAccount = await wallet.getAccount(process.env.FACTORY_CONTRACT)
+              
+              let appClient = await ceramic.getCeramic(appAccount, appSeed)
+
+              let contract = await dao.loadDAO(contractId)
+              setContract(contract)
+              console.log('dao contract', contract)
+
+              let init = await contract.getInit()
+              console.log('init', init)
+
+              let thisDIDsContract = await dids.loadDIDs(process.env.DIDS_CONTRACT)
+              setDidsContract(thisDIDsContract)
+
+              let thisContractIdx
+              let currentAliases = {}
+              try {
+                  let allAliases = await thisDIDsContract.getAliases()
+                 
+                  //reconstruct aliases
+                  let i = 0
+                  
+                  while (i < allAliases.length) {
+                      let key = allAliases[i].split(':')
+                      let alias = {[key[0]]: key[1]}
+                      currentAliases = {...currentAliases, ...alias}
+                      i++
+                  }
+                  if(allAliases) {
+                      thisContractIdx = new IDX({ ceramic: appClient, aliases: currentAliases})
+                      setContractIdx(thisContractIdx)
+                      handleContractIdx(thisContractIdx)
+                  }
+              } catch (err) {
+                  console.log('error retrieving aliases and setting app Idx', err)
+              }
+              console.log('contractidx', thisContractIdx)
+
+              // Set Current User Ceramic Client
+              console.log('accountid', accountId)
+                   
+              let thisUserIdx = new IDX({ ceramic: thisCurrentUserCeramicClient, aliases: currentAliases})
+              setCurUserIdx(thisUserIdx)              
+            
+              let did = await thisDIDsContract.getDID({accountId: contractId})
+              console.log('did', did)
+
+              let memberEvents = await thisContractIdx.get('member', did)
+              console.log('memberEvents', memberEvents)
+              setAllMemberInfo(memberEvents.events)
+
+              let memberProposalEvents = await thisContractIdx.get('memberProposal')
+              console.log('memberproposalevents', memberProposalEvents)
+              if(memberProposalEvents && memberProposalEvents.events.length > 0){
+              console.log('memberProposalEvents', memberProposalEvents)
+              setProposalEvents(memberProposalEvents.events)
+              }
+
+              let thisMemberInfo
+              let thisMemberStatus
+
+              try {
+                thisMemberStatus = await contract.getMemberStatus({member: accountId})
+                console.log('member status', thisMemberStatus)
+                if(isMounted) {
+                setMemberStatus(thisMemberStatus)
+                }
+              } catch (err) {
+                console.log('no member status yet')
+               
+              }
+             
+              try {
+                thisMemberInfo = await contract.getMemberInfo({member: accountId})
+                console.log('member info', thisMemberInfo)
+                if(isMounted) {
+                setMemberInfo(thisMemberInfo)
+                }
+              } catch (err) {
+                console.log('no member info yet')
+               
+              }
+
+              try {
+                let owner = await contract.getSummoner()
+                if(isMounted) {
+                setSummoner(owner)
+                console.log('summoner', summoner)
+                }
+              } catch (err) {
+                console.log('no summoner yet')
+               
+              }
+  
+              try {
+                let shares = await contract.getTotalShares()
+                console.log('shares', shares)
+                if(isMounted) {
+                setTotalShares(shares)
+                }
+              } catch (err) {
+                console.log('no total shares yet')
+               
+              }
+
+              try {
+                let token = await contract.getDepositToken()
+                if(isMounted) {
+                setDepositToken(token)
+                }
+              } catch (err) {
+                console.log('no deposit token yet')
+               
+              }
+                  
+              try {
+                let deposit = await contract.getProposalDeposit()
+                console.log('proposal deposit', deposit)
+                if(isMounted) {
+                setProposalDeposit(deposit)
+                }
+              } catch (err) {
+                console.log('no proposal deposit yet')
+               
+              }
+  
+              try {
+                let duration = await contract.getPeriodDuration()
+                if(isMounted) {
+                setPeriodDuration(duration)
+                }
+              } catch (err) {
+                console.log('no period duration yet')
+               
+              }
+
+              let ebalance
+              try {
+                ebalance = await contract.getEscrowTokenBalances()
+                console.log('escrow balance', ebalance)
+                if(isMounted) {
+                setEscrowBalance(ebalance)
+                }
+              } catch (err) {
+                console.log('no escrow balance')
+               
+              }
+
+              let gbalance
+              try {
+                gbalance = await contract.getGuildTokenBalances()
+                if(isMounted) {
+                setGuildBalance(gbalance)
+                }
+              } catch (err) {
+                console.log('no guild balance')
+              
+              }
+
+              if(thisMemberStatus && thisMemberInfo !== undefined) {
               if(isMounted) {
                 setMemberIcon(<CheckCircleIcon />)
-                setSharesLabel('Shares: ' + memberInfo[0].shares)
-                setLootLabel('Loot: ' + memberInfo[0].loot)
+                setSharesLabel('Shares: ' + thisMemberInfo[0].shares)
+                setLootLabel('Loot: ' + thisMemberInfo[0].loot)
               }
               }
 
               let guildRow
-                if(guildBalance) {
-                  for (let i = 0; i < guildBalance.length; i++) {
-                    guildRow = (<>{guildBalance[i].balance} {guildBalance[i].token}</>
+                if(gbalance) {
+                  for (let i = 0; i < gbalance.length; i++) {
+                    guildRow = (<>{gbalance[i].balance} {gbalance[i].token}</>
                     )
                   }
                 } else {
                   guildRow = '0 Ⓝ'
                 }
+
                 if(isMounted){
                 setGuildBalanceChip(<>{guildRow}</>)
                 }
 
               let escrowRow
-                if(escrowBalance) {
-                  for (let i = 0; i < escrowBalance.length; i++) {
-                    escrowRow = (<>{escrowBalance[i].balance} {escrowBalance[i].token}</>)
+                if(ebalance) {
+                  for (let i = 0; i < ebalance.length; i++) {
+                    escrowRow = (<>{ebalance[i].balance} {ebalance[i].token}</>)
                   }
                 } else {
                   escrowRow = '0 Ⓝ'
                 }
+
                 if(isMounted) {
                 setEscrowBalanceChip(<>{escrowRow}</>)
                 }
 
               let getNearPrice = await axios.get('https://api.coingecko.com/api/v3/simple/price?ids=near&vs_currencies=usd')
+              console.log('nearprice', getNearPrice.data.near.usd)
               if(isMounted){
                 setNearPrice(getNearPrice.data.near.usd)
               }
+
+
+              if(currentPeriod == undefined || currentPeriod == 0){
+                try {
+                  let period = await contract.getCurrentPeriod()
+                  if(isMounted) {
+                    setCurrentPeriod(period)
+                  }
+                } catch (err) {
+                  console.log('get period issue', err)
+                }
+              }
+
+              let i = 1
+              setTimeout(async function refreshCurrentPeriod() {
+                let start = true
+                let init
+                try{
+                    init = await contract.getInit()
+                } catch (err) {
+                    console.log('cant retreive init', err)
+                }
+                if(init=='done'){
+                try {
+                let period = await contract.getCurrentPeriod()
+                setCurrentPeriod(period)
+                console.log('get period success')
+                } catch (err) {
+                console.log('get period issue', err)
+                }
+                start = false
+                i++
+                if(start == false){
+                setTimeout(refreshCurrentPeriod, 10000)
+                }
+            }
+            }, 10000)
             }
 
             fetchData()
@@ -184,12 +478,17 @@ export default function AppFramework(props) {
                
               })
               return () => { isMounted = false } // use effect cleanup to set flag false if unmounted
-        }, [escrowBalance, guildBalance]
+        }, [ memberStatus ]
       )
+
+    function handleTabValueState(value) {
+      setTabValue(value)
+    }
+    
   
     return (
             <>
-            <Grid container style={{padding:'5px'}}>
+            <Grid container style={{padding:'20px'}}>
             <Grid container justify="space-evenly" alignItems="center" style={{marginBottom:'15px'}} spacing={0}>
               <Grid item xs={12} sm={12} md={6} lg={6} xl={6}>
               <div style={{marginLeft: '10px'}}>
@@ -206,12 +505,24 @@ export default function AppFramework(props) {
                   contract={contract}
                   daoContract={daoContract}
                   proposalDeposit={proposalDeposit}
+                  didsContract={didsContract}
+                  contractIdx={contractIdx}
+                  curUserIdx={curUserIdx}
                 />
               </div>
               </Grid>
               <Grid item xs={12} sm={12} md={6} lg={6} xl={6}>
               <div style={{float:'right'}}>
-              {summoner == accountId ?<div style={{float:'right',marginTop:'-5px',marginLeft:'5px'}}><RightSideDrawer handleInitChange={handleInitChange} accountId={accountId} contract={contract}/></div>: null }
+              {summoner == accountId ?<div style={{float:'right',marginTop:'-5px',marginLeft:'5px'}}><RightSideDrawer 
+                handleInitChange={handleInitChange} 
+                accountId={accountId} 
+                contract={contract}
+                hasDao={hasDao}
+                factoryContract={factoryContract} 
+                handleHasDao={handleHasDao} 
+                handleErrorMessage={handleErrorMessage} 
+                handleSuccessMessage={handleSuccessMessage}
+                handleSnackBarOpen={handleSnackBarOpen} /></div>: null }
                 <Chip variant="outlined" label="Member" icon={memberIcon} />
                 <Chip variant="outlined" label={lootLabel}  />
                 <Chip variant="outlined" label={sharesLabel}  />
@@ -223,16 +534,16 @@ export default function AppFramework(props) {
         
           <Grid container justify="center" alignItems="center" spacing={1} className={classes.top}> 
             <Grid item xs={12} sm={12} md={3} lg={3} xl={3}>
-              <Typography variant="h6" color="textPrimary" align="center">Fund: {guildBalanceChip} {guildBalance.length > 0 ? guildBalance[0].balance > 0 ? '($' + (parseInt(guildBalance[0].balance) * nearPrice).toFixed(2) + ' USD)' : '($0.00 USD)' : null } </Typography>
+              <Typography variant="h6" color="textPrimary" align="center">Fund: {guildBalanceChip} {guildBalance && guildBalance.length > 0 ? guildBalance[0].balance > 0 ? '($' + (parseInt(guildBalance[0].balance) * nearPrice).toFixed(2) + ' USD)' : '($0.00 USD)' : null } </Typography>
             </Grid>
             <Grid item xs={12} sm={12} md={3} lg={3} xl={3}>
-              <Typography variant="h6" color="textPrimary" align="center">Escrow: {escrowBalanceChip} {escrowBalance.length > 0 ? escrowBalance[0].balance > 0 ? '($' + (parseInt(escrowBalance[0].balance) * nearPrice).toFixed(2) + ' USD)' : '($0.00 USD)' : null }</Typography>
+              <Typography variant="h6" color="textPrimary" align="center">Escrow: {escrowBalanceChip} {escrowBalance && escrowBalance.length > 0 ? escrowBalance[0].balance > 0 ? '($' + (parseInt(escrowBalance[0].balance) * nearPrice).toFixed(2) + ' USD)' : '($0.00 USD)' : null }</Typography>
             </Grid>
             <Grid item xs={12} sm={12} md={3} lg={3} xl={3}>
               <Typography variant="h6" color="textPrimary" align="center">Total Shares: {totalShares}</Typography>
             </Grid>
             <Grid item xs={12} sm={12} md={2} lg={2} xl={2}>
-              <Typography variant="h6" color="textPrimary" align="center">Share Value: {guildBalance.length > 0 ? guildBalance[0].balance > 0 ? '$' + ((guildBalance[0].balance/totalShares)*nearPrice).toFixed(2) + ' USD' : '$0.00 USD' : null }</Typography>
+              <Typography variant="h6" color="textPrimary" align="center">Share Value: {guildBalance && guildBalance.length > 0 ? guildBalance[0].balance > 0 ? '$' + ((guildBalance[0].balance/totalShares)*nearPrice).toFixed(2) + ' USD' : '$0.00 USD' : null }</Typography>
             </Grid>
           </Grid>
           
@@ -264,6 +575,16 @@ export default function AppFramework(props) {
                 contract={contract}
                 daoContract={daoContract}
                 allMemberInfo={allMemberInfo}
+                handleSnackBarOpen={handleSnackBarOpen}
+                handleSuccessMessage={handleSuccessMessage}
+                handleErrorMessage={handleErrorMessage}
+                summoner={summoner}
+                contractIdx={contractIdx}
+                curUserIdx={curUserIdx}
+                didsContract={didsContract}
+                contractId={contractId}
+                appIdx={appIdx}
+                appClient={appClient}
               />
             </Grid>
           </Grid>
