@@ -2,10 +2,12 @@ import React from 'react';
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
 //import NearLogo from '../svg/NearLogo';
-import FleetsLogo from '../../images/fleets-black.png'
+import CatalystLogo from '../../images/catalyst-logo-cropped.png'
 
 const StyledLogo = styled(Link)`
     margin-top: 5px;
+    float:left;
+    margin-left: 20px;
 
     @media (max-width: 991px) {
         max-width: 53px;
@@ -17,11 +19,15 @@ const StyledLogo = styled(Link)`
     svg {
         width: 155px;
     }
+
+    img {
+        width: 200px;
+    }
 `
 
 const Logo = () => (
     <StyledLogo to='/' className='logo'>
-        <img src={FleetsLogo} alt="Fleets Logo"/>
+        <img src={CatalystLogo} alt="Catalyst Logo"/>
     </StyledLogo>
 )
 
